+++ src/Products/MeetingCommunes/adapters.py
@@ -853,14 +853,6 @@ class MeetingCommunesWorkflowConditions(MeetingWorkflowConditions):
     implements(IMeetingCommunesWorkflowConditions)
     security = ClassSecurityInfo()
 
-    security.declarePublic('mayCorrect')
-
-    def mayCorrect(self, destinationState=None):
-        '''Override to avoid call to _decisionsWereConfirmed.'''
-        if not _checkPermission(ReviewPortalContent, self.context):
-            return
-        return True
-
     security.declarePublic('mayDecide')
 
     def mayDecide(self):